requirements.txt
pybit>=5.0.0
python-telegram-bot>=20.0
pandas
numpy
flask
gspread>=6.0.0
oauth2client>=4.1.3
python-dotenv
